+++ webui/src/models/cluster/server-suggestions/selectors.ts
@@ -1,0 +1,7 @@
+import type { GetCompressionCluster, GetCompressionClusterCompressionCompressionInfo } from './types';
+
+export const clusterCompressionInfo = (
+  data: GetCompressionCluster
+): GetCompressionClusterCompressionCompressionInfo => {
+  return data?.cluster?.cluster_compression.compression_info ?? [];
+};